backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.51.0
python-dotenv==1.0.1
rapidfuzz==3.9.7
pydantic==2.9.2
